requirements.txt
# You can use the next command to install all the necessary modules:
# `pip install -r requirements.txt`

matplotlib==3.6.0
networkx==2.8.6
numpy==1.23.3
jpype1==1.5.0
ffmpeg==1.4
pandas==2.0.0
openpyxl==3.1.2
textdistance==4.6.1
termcolor==2.4.0
PyQt5==5.15.9
pyqt5-tools~=5.15
scipy==1.12.0
pydantic==2.6.3
pydantic_settings==2.2.1
sqlalchemy==2.0.28
psycopg2==2.9.9
fastapi==0.110.0
uvicorn==0.28.0
